Adding samples Moved GISUtils to solve class loader issues. Modified Geometry Information step to use Geometry Interface to avoid class loader issues.
