+++ pentaho-gis-plugins/src/main/java/com/atolcd/pentaho/di/gis/io/DXFReader.java
@@ -32,7 +32,7 @@ import com.atolcd.gis.dxf.Layer;
 import com.atolcd.pentaho.di.gis.io.features.Feature;
 import com.atolcd.pentaho.di.gis.io.features.Field;
 import com.atolcd.pentaho.di.gis.io.features.Field.FieldType;
-import com.atolcd.pentaho.di.gis.utils.GeometryUtils;
+import com.atolcd.pentaho.di.utils.GeometryUtils;
 
 public class DXFReader extends AbstractFileReader {
