+++ pentaho-gis-plugins/src/main/java/com/atolcd/pentaho/di/gis/io/DXFWriter.java
@@ -35,7 +35,7 @@ import java.util.regex.Pattern;
 import org.pentaho.di.core.exception.KettleException;
 
 import com.atolcd.pentaho.di.gis.io.features.Feature;
-import com.atolcd.pentaho.di.gis.utils.GeometryUtils;
+import com.atolcd.pentaho.di.utils.GeometryUtils;
 import com.vividsolutions.jts.geom.Geometry;
 
 import fr.michaelm.jump.drivers.dxf.DxfFile;
