+++ pentaho-gis-plugins/src/main/java/com/atolcd/pentaho/di/gis/io/GeoJSONReader.java
@@ -40,7 +40,7 @@ import org.wololo.geojson.GeoJSONFactory;
 import com.atolcd.pentaho.di.gis.io.features.Feature;
 import com.atolcd.pentaho.di.gis.io.features.Field;
 import com.atolcd.pentaho.di.gis.io.features.Field.FieldType;
-import com.atolcd.pentaho.di.gis.utils.GeometryUtils;
+import com.atolcd.pentaho.di.utils.GeometryUtils;
 import com.vividsolutions.jts.geom.Geometry;
 
 public class GeoJSONReader extends AbstractFileReader {
