+++ pentaho-gis-plugins/src/main/java/com/atolcd/pentaho/di/gis/io/MapInfoReader.java
@@ -35,7 +35,7 @@ import org.pentaho.di.core.exception.KettleException;
 import com.atolcd.pentaho.di.gis.io.features.Feature;
 import com.atolcd.pentaho.di.gis.io.features.Field;
 import com.atolcd.pentaho.di.gis.io.features.Field.FieldType;
-import com.atolcd.pentaho.di.gis.utils.GeometryUtils;
+import com.atolcd.pentaho.di.utils.GeometryUtils;
 import com.vividsolutions.jts.geom.Geometry;
 
 public class MapInfoReader extends AbstractFileReader {
