+++ pentaho-gis-plugins/src/main/java/com/atolcd/pentaho/di/trans/steps/giscoordinatetransformation/GisCoordinateTransformation.java
@@ -48,8 +48,8 @@ import org.pentaho.di.trans.step.StepMeta;
 import org.pentaho.di.trans.step.StepMetaInterface;
 
 import com.atolcd.pentaho.di.core.row.value.ValueMetaGeometry;
-import com.atolcd.pentaho.di.gis.utils.CoordinateTransformer;
-import com.atolcd.pentaho.di.gis.utils.GeometryUtils;
+import com.atolcd.pentaho.di.utils.CoordinateTransformer;
+import com.atolcd.pentaho.di.utils.GeometryUtils;
 import com.vividsolutions.jts.geom.Geometry;
 
 public class GisCoordinateTransformation extends BaseStep implements StepInterface {
