+++ samples/transformations/read-geometry/read-csv/geom.csv
@@ -1,0 +1,2 @@
+GEOMETRY,ID
+POINT(0 1),1
